feed/requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.10.8
django-haystack==2.5.1
elasticsearch==5.0.1
feedparser==5.2.1
idna==2.6
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
schedule==0.4.3
urllib3==1.22
